Introduce Utility to Deploy From GitHub Issue
This PR adds a new `github-deploy.sh` script that can be used to deploy
the Safe singleton factory from a GitHub issue. It also supports both
multiple GitHub issue numbers being input at a time, or with empty
parameters defaults to deploying for all GitHub issues that are marked
with the `ready-to-deploy` tag.