Add Script for Verifying Deployments
This PR adds a new `yarn verify` script that checks deployed bytecode
from the RPC configuration. This makes it easier to review PRs for new
networks.

As a pro-tip, instead of using a `.env` file and setting the `RPC` in
there, you can just invoke the command:

```
RPC=https://example.com yarn verify
```

(Note that the `RPC` variable in the `.env` file **will take
precedence**, so make sure to either remove the `.env` file or not set
the `RPC` variable in it when using this trick).

Note that this **does not support zkSync based networks at the moment**.