[New Chain] Chronos zkEVM Testnet (#483)
Fixes #480

This PR deploys the singleton factory to the Chronos test network. Note
that this is a ZK test network, so the automated test scripts will not
work.

In addition, some adjustments to the `hardhat` configuration were
required to get this deploying. It looks like `hardhat` + `zksync-web3`
are always trying to connect to a network from the hardhat configuration
(regardless of whether or not it is used, which it shouldn't be in this
case since we are specifying an RPC for the provider to connect to).
This would cause network errors as the default zkSync network to connect
to was the Goerli testnet which is deprecated, and its RPC is no longer
accessible. With that in mind, in order to work around the issue, we
just needed to update the configuration to point to the lastest Sepolia
test network.

Contract was deployed to
[`0xaECDbB0a3B1C6D1Fe1755866e330D82eC81fD4FD`](https://zkevm-t0.cronos.org/explorer/address/0xaECDbB0a3B1C6D1Fe1755866e330D82eC81fD4FD#contract).
You can verify this matches the deployment on [zkSync
mainnet](https://explorer.zksync.io/address/0xaECDbB0a3B1C6D1Fe1755866e330D82eC81fD4FD#contract).

> Note: At the moment, the explorer does not seem to be indexing new
transactions, so the above link may not be working. You can use the RPC
in the linked issue to query the blockchain directly and verify the
code.