Check Deployment Simulation
This PR changes the automated script to also check for contract creation
transaction simulation works as expected. the `npm run submit` script
requires this to work, and so we check it to prevent spending additional
time reporting this issue when trying to deploy the contracts.

Tested with:

```sh
bash .github/scripts/validate_new_chain_request.sh --debug 765
```