Fix New Chain CI
It looks like the `SUMMARY_FILE` was only set for the `npm` command and
not for the `jq` commands where the outputs were read. This lead to
`COMMENT_OUTPUT` and `LABEL_OPERATION` being empty and causing issues in
the CI step that posts the result to the GitHub issue.

Additionally, this PR does some minor formatting changes and modifies
the error messages for already deployed factories to be less scary and
notify the user that PRs for the artifacts are expected.

Tested the CI locally with:

```
$ SUMMARY_FILE=summary.json
$ GITHUB_ENV=github.env
$ ISSUE_BODY="$(gh issue view 1004 --json body --jq '.body')"
$ RPC="$(echo "$ISSUE_BODY" | grep -E -o 'https?://[^ ]+' -m 1 | head -1)"\
    npm run verify:new-chain-request || true

> @safe-global/safe-singleton-factory@1.0.41 verify:new-chain-request
> ts-node scripts/new-chain.ts

{ chainId: 49321 }
{
  chainlist: 'https://raw.githubusercontent.com/ethereum-lists/chains/master/_data/chains/eip155-49321.json',
  onChainlist: true
}
{
  nonce: 0,
  codehash: '0xc5d2460186f7233c927e7db2dcc703c0e500b653ca82273b7bfad8045d85a470',
  code: '0x'
}
{
  gasPrice: '100000001',
  gasLimit: '68990',
  gasEstimate: '10348500103485'
}
{
  simulation: '0x7fffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffe03601600081602082378035828234f58015156039578182fd5b8082525050506014600cf3'
}
{
  simulationCodehash: '0x2fa86add0aed31f33a762c9d88e807c475bd51d0f52bd0955754b2608f7e4989'
}
{ balance: '10000000000000000000' }
{
  commentOutput: '**✅ Success:**<br>The issue description is valid:<br>- The RPC URL is valid<br>- The chain is in the chainlist<br>- The deployer address is pre-funded<br>:sparkles: The team will be in touch with you soon :sparkles:',
  labelOperation: '--add-label'
}
$ echo "COMMENT_OUTPUT=$(jq -r '.commentOutput' $SUMMARY_FILE)" >> $GITHUB_ENV
$ echo "LABEL_OPERATION=$(jq -r '.labelOperation' $SUMMARY_FILE)" >> $GITHUB_ENV
$ cat $GITHUB_ENV
COMMENT_OUTPUT=**✅ Success:**<br>The issue description is valid:<br>- The RPC URL is valid<br>- The chain is in the chainlist<br>- The deployer address is pre-funded<br>:sparkles: The team will be in touch with you soon :sparkles:
LABEL_OPERATION=--add-label
```