Switch to EVM Bytecode Interpreter on ZKsync Networks
We now use the EVM bytecode version of the Safe singleton factory by
default for the ZKsync based networks that we officially support in
Safe. This will make it so Save v1.5.0 will correctly deploy EVM
versions of the contract.

Note that I `s/zkSync/ZKsync/g`, as they changed their logo and
capilization on their website, so I thought it better reflect that in
our documentation here as well.