Add Ledger Support (#1093)
This PR adds support for the Ledger wallet for signing Singleton factory
deployments. It also soft-deprecates the `MNEMONIC` method (and displays
a warning in the console when used) in order to promote better security
hygiene.

I tested both EVM and ZKsync deployments, and both seemed to work.

Note that we need to do some monkey-patching on the `LedgerSigner`
implementation for our use-case. See the comment in the
`LedgerTypedDataSigner` class for more information.