CI Improvements
I noticed when developing #1096 that we don't actually typecheck our
scripts in CI. This PR adds some adds a new `tsconfig.dev.json`
configuration that does not emit code, and lets us typecheck all our
sources in CI.

This found a typing error in `hardhat.config.ts`, where `verifierURL`
is not a known property. I cleaned up the hardhat configuration a
bit (since the network configurations aren't used or needed, and the
compiler version is in fact fixed to a historic version, and should
not be modified to use the latest available version).