Parse `v`-Value from Ledger Device (#1136)
The Ledger device API returns the signature `v` value as a hexadecimal
encoded string and **not** a number. This causes unexpected behaviour
with the bitwise operations, and why the `yParity` computation would
sometimes fail.

This PR fixes the issue by first `parseInt`-ing the `v` value in base 16
(i.e. hexadecimal) before applying the bitwise operations.